Added option to delete tasks

+++ src/AppBundle/Controller/TaskController.php
@@ -4,6 +4,7 @@ namespace AppBundle\Controller;
 
 use AppBundle\Entity\Task;
 use AppBundle\Form\TaskType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -73,7 +74,7 @@ class TaskController extends Controller
      * @param $id
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function showTaskAction($id) {
+    public function showTaskAction($id, Request $request) {
         $task = $this->getDoctrine()
                     ->getRepository('AppBundle:Task')
                     ->find($id);
@@ -105,18 +106,40 @@ class TaskController extends Controller
     /**
      * @Route("/delete/{id}", name="delete_task")
      *
+     * @param Request $request
      * @param $id
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
-    public function removeTaskAction($id) {
-        $task = $this->getDoctrine()
-            ->getRepository('AppBundle:Task')
-            ->find($id);
+    public function deleteTaskAction(Request $request, $id) {
+        $form = $this->createDeleteForm($id);
 
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
 
+        if($request->isMethod('POST')) {
+            $form->submit($request->request->get($form->getName()));
+
+
+            if ($form->isValid()) {
+                $task = $this->getDoctrine()
+                    ->getRepository('AppBundle:Task')
+                    ->find($id);
+
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($task);
+                $em->flush();
+
+            }
+        }
         return $this->redirectToRoute('home');
+
+    }
+
+    private function createDeleteForm($taskId) {
+        return $this->createFormBuilder(array('id' => $taskId))
+            ->add('id', TextType::class,
+                array(
+                    'attr' => array('hidden' => true)))
+            ->setAction($this->generateUrl('delete_task', array('id' => $taskId)))
+            ->setMethod('POST')
+            ->getForm();
     }
 }
